requirements.txt
cycler==0.11.0
fonttools==4.30.0
kiwisolver==1.3.2
matplotlib==3.5.1
minimalmodbus==2.0.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
